피드백 수정 및 엔티티 클래스 추가, 중복 칼럼 관련 클래스 추가

+++ src/main/java/com/example/record/RecordRepository.java
@@ -2,5 +2,5 @@ package com.example.record;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecordRepository extends JpaRepository<Record, Integer> {
+public interface RecordRepository extends JpaRepository<Record, Long> {
 }
